test(e2e): rename the geolocation helper to stubLocation
Reviewer-suggested name; the file stubs a location rather than
providing geolocation.
